[#9] Move changing password form on separate account page
