+++ views/index.ejs
@@ -19,6 +19,9 @@
                     <a class="nav-link" href="/panel">Panel</a>
                 </li>
                 <li class="nav-item">
+                    <a class="nav-link" href="/account">Account</a>
+                </li>
+                <li class="nav-item">
                     <a class="nav-link" href="/logout">Logout</a>
                 </li>
             <% } else { %>
