+++ views/panel/new.ejs
@@ -18,6 +18,9 @@
                 <a class="nav-link active" href="/panel">Panel</a>
             </li>
             <li class="nav-item">
+                <a class="nav-link" href="/account">Account</a>
+            </li>
+            <li class="nav-item">
                 <a class="nav-link" href="/logout">Logout</a>
             </li>
         </ul>
